add event aliases for primary events
for those events that depend on terror level, invader stage, or
healthy/blighted island, now it will be possible to look up an event by
the second option. This is useful for someone who wants to look up e.g.
Increasing Aggression without having to remember which stage I event
it's associated with.